BDcrud3/requirements.txt
fastapi
uvicorn[standard]
pydantic>=2
python-dotenv
psycopg2-binary
pymongo
